Change files modes and rename files
The execution permission was removed from `Team_1_proposal.latex' and
`IEEEtran.cls'. These are not executable files, and this bit should not
be set.

The file `/sections/Pictures/image.png' was also renamed to
`/images/architecture-diagram.png'. This necessitated changes in
`/Team_1_proposal.latex' and `/sections/system-and-functionality.latex'
for the new path and file name.